# beacon-server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "MeshCore network observation backend. Ingests LoRa packets from MQTT brokers, stores in PostgreSQL, and streams live events via WebSocket.",
        "title": "MeshCore Beacon API",
        "termsOfService": "https://github.com/MeshCore-Beacon/beacon-server",
        "contact": {
            "name": "MeshCore Beacon",
            "url": "https://github.com/MeshCore-Beacon/beacon-server"
        },
        "license": {
            "name": "AGPL-3-or-later"
        },
        "version": "1.6.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/observers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Observers"
                ],
                "summary": "List observers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by single IATA code (case-insensitive)",
                        "name": "iata",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by multiple IATA codes, comma-separated e.g. YVR,YYJ",
                        "name": "iatas",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by region ID, expands to member IATAs",
                        "name": "regionId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by region slug, expands to member IATAs",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by observer type (e.g. meshcoretomqtt, meshcore-ha)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by broker name",
                        "name": "broker",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (online or offline)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Partial case-insensitive display name match",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by transport scope name e.g. %23bc (URL-encoded #bc)",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field: name, type, radio, iata, status, last_seen (default last_seen)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort direction: asc or desc (default desc)",
                        "name": "direction",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Opaque keyset cursor returned as nextPageToken",
                        "name": "pageToken",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Legacy last_seen epoch ms cursor (only for last_seen desc)",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_ObserverSummary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
        "/observers/{observerId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Observers"
                ],
                "summary": "Get observer detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Observer UUID",
                        "name": "observerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Observer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
        "/observers/{observerId}/adverts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Observers"
                ],
                "summary": "List advert packets heard by an observer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Observer UUID",
                        "name": "observerId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Observation ID of last item for pagination",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_AdvertObservation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
        "/observers/{observerId}/telemetry": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Observers"
                ],
                "summary": "Get observer telemetry history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Observer UUID",
                        "name": "observerId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Duration window e.g. 24h, 48h, 168h (default 24h)",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Return points after this telemetry ID for WS reconnection backfill",
                        "name": "afterId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bucketing interval: 1h (default), 6h, or 24h",
                        "name": "interval",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverTelemetry"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_MeshCore-Beacon_beacon-server_internal_api.AdvertObservation": {
            "type": "object",
            "properties": {
                "heardAt": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "hopCount": {
                    "type": "integer"
                },
                "iata": {
                    "type": "string"
                },
                "id": {
                    "description": "observation ID, use as cursor for pagination",
                    "type": "integer"
                },
                "nodeName": {
                    "type": "string"
                },
                "nodePublicKey": {
                    "description": "hex-encoded",
                    "type": "string"
                },
                "packetHash": {
                    "description": "hex-encoded",
                    "type": "string"
                },
                "payloadType": {
                    "type": "integer"
                },
                "payloadTypeName": {
                    "type": "string"
                },
                "rssi": {
                    "type": "integer"
                },
                "snr": {
                    "type": "number"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.Observer": {
            "type": "object",
            "properties": {
                "batteryLevel": {
                    "description": "volts, nil if mains powered",
                    "type": "number"
                },
                "brokers": {
                    "description": "broker names this observer has been seen on",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverBroker"
                    }
                },
                "displayName": {
                    "description": "friendly name from /status messages",
                    "type": "string"
                },
                "firmwareBuild": {
                    "type": "string"
                },
                "firmwareVersion": {
                    "type": "string"
                },
                "firstSeen": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "hardwareModel": {
                    "type": "string"
                },
                "iata": {
                    "description": "most recently heard IATA",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastSeen": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "lastStatusAt": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "observationCount": {
                    "type": "integer"
                },
                "observerType": {
                    "description": "e.g. \"meshcoretomqtt\", \"meshcoreha\"",
                    "type": "string"
                },
                "publicKey": {
                    "description": "hex-encoded public key",
                    "type": "string"
                },
                "radio": {
                    "description": "friendly radio param string: freqMhz,BwKhz,SF e.g. \"910.525,62.5,7\"",
                    "type": "string"
                },
                "radioBwKhz": {
                    "description": "bandwidth in kHz",
                    "type": "number"
                },
                "radioCr": {
                    "description": "coding rate denominator",
                    "type": "integer"
                },
                "radioFreqMhz": {
                    "description": "MHz e.g. 910.525",
                    "type": "number"
                },
                "radioSf": {
                    "description": "LoRa spreading factor",
                    "type": "integer"
                },
                "scopes": {
                    "description": "list of observer forwarded scopes matched to config",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "softwareVersion": {
                    "type": "string"
                },
                "status": {
                    "description": "\"online\" or \"offline\" derived from last_status_at",
                    "type": "string"
                },
                "statusMetadata": {
                    "description": "raw /status JSON payload"
                },
                "uptimeSeconds": {
                    "type": "integer"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverBroker": {
            "type": "object",
            "properties": {
                "lastPacketAt": {
                    "description": "epoch ms, last packet received via this broker; 0 if none",
                    "type": "integer"
                },
                "lastSeenAt": {
                    "description": "epoch ms, last time observer was seen on this broker",
                    "type": "integer"
                },
                "name": {
                    "description": "broker name e.g. \"mqtt1\"",
                    "type": "string"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverSummary": {
            "type": "object",
            "properties": {
                "displayName": {
                    "description": "friendly name from /status messages",
                    "type": "string"
                },
                "iata": {
                    "description": "most recently heard IATA",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "observerType": {
                    "description": "e.g. \"meshcoretomqtt\", \"meshcoreha\"",
                    "type": "string"
                },
                "radio": {
                    "description": "friendly radio param string: freqMhz,BwKhz,SF e.g. \"910.525,62.5,7\"",
                    "type": "string"
                },
                "scopes": {
                    "description": "list of observer forwarded scopes matched to config",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "description": "\"online\" or \"offline\" derived from last_status_at",
                    "type": "string"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverTelemetry": {
            "type": "object",
            "properties": {
                "interval": {
                    "type": "string"
                },
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverTelemetryPoint"
                    }
                },
                "range": {
                    "type": "string"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverTelemetryPoint": {
            "type": "object",
            "properties": {
                "airtimeRxPct": {
                    "type": "number"
                },
                "airtimeTxPct": {
                    "type": "number"
                },
                "batteryMv": {
                    "type": "integer"
                },
                "noiseFloorDb": {
                    "type": "number"
                },
                "queueLength": {
                    "type": "integer"
                },
                "receiveErrors": {
                    "type": "integer"
                },
                "t": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "uptimeSeconds": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_AdvertObservation": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.AdvertObservation"
                    }
                },
                "nextCursor": {
                    "type": "integer"
                },
                "nextPageToken": {
                    "type": "string"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_ObserverSummary": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ObserverSummary"
                    }
                },
                "nextCursor": {
                    "type": "integer"
                },
                "nextPageToken": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "e.g. \"not_found\", \"bad_request\"",
                    "type": "string"
                },
                "message": {
                    "description": "e.g. \"channel not found\"",
                    "type": "string"
                }
            }
        },
...
    },
...
}